task1/Practical Work/Repo_03_LIteLLM_Openai_agent/02_litellm_uv_chainlit/chatbot/requirements.txt
chainlit
litellm
python-dotenv